requirements.txt
opencv-python>=4.8.0
# 0.10.32 fixes "function 'free' not found" on Windows (see google-ai-edge/mediapipe#6187)
mediapipe>=0.10.32
# Windows: helps MediaPipe load; harmless on other platforms
msvc-runtime; sys_platform == "win32"
speechrecognition>=3.10.0
PyAudio>=0.2.14
Pillow>=10.0.0
numpy>=1.24.0
vosk>=0.3.45

# Optional: better accuracy than Vosk (pip install faster-whisper)
# faster-whisper

# Optional: live translation (press M in app)
# googletrans==4.0.0-rc1

# Optional: load .env for DEEPGRAM_API_KEY etc.
python-dotenv>=1.0.0

# Optional: best quality cloud STT (set DEEPGRAM_API_KEY in .env)
deepgram-sdk>=5.0.0
